fix: move BeaconBlockField to the host crate
