+++ crates/host-executor/src/lib.rs
@@ -1,13 +1,14 @@
 //! SP1 Contract Call Host Executor Library
 //!
-//! This library provides functionality for executing contract calls within the SP1 zkVM host environment.
+//! This library provides functionality for executing contract calls within the SP1 zkVM host
+//! environment.
 
 pub use rsp_primitives::genesis::Genesis;
 
 mod anchor_builder;
 pub use anchor_builder::{
-    AnchorBuilder, BeaconAnchorBuilder, BeaconAnchorKind, ChainedBeaconAnchorBuilder,
-    ConsensusBeaconAnchor, Eip4788BeaconAnchor, HeaderAnchorBuilder,
+    AnchorBuilder, BeaconAnchorBuilder, BeaconAnchorKind, BeaconBlockField,
+    ChainedBeaconAnchorBuilder, ConsensusBeaconAnchor, Eip4788BeaconAnchor, HeaderAnchorBuilder,
 };
 
 mod beacon;
